2026/day-36/requirements.txt
flask 
mysql-connector-python
python-dotenv

